--- src/main/app.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import AppRouter from './router';
-
-export default class App extends React.Component {
-  render() {
-    return <AppRouter />;
-  }
-}
