--- src/main/main.jsx
@@ -1,62 +1,0 @@
-import React from 'react';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-
-import Header from './header';
-import MainMenu from './menu';
-import LoginForm from '../loginform';
-
-export default class Main extends React.Component {
-  componentWillMount() {
-    this.closeMainMenu();
-  }
-
-  openMainMenu() {
-    this.setState({mainMenu: true});
-  }
-
-  closeMainMenu() {
-    this.setState({mainMenu: false});
-  }
-
-  render() {
-    let children = this.props.children;
-    let dialog = children.type.dialog || false;
-    let closable = children.type.closable || true;
-    if (typeof dialog === 'function') {
-      dialog = dialog(this.props.params);
-    }
-    if (typeof closable === 'function') {
-      closable = closable(this.props.params);
-    }
-    return (
-      <MuiThemeProvider>
-        <div>
-          <Header
-            dialog={dialog}
-            closable={closable}
-            openMainMenu={this.openMainMenu.bind(this)}
-          />
-          <MainMenu
-            enable={this.state.mainMenu}
-            open={this.openMainMenu.bind(this)}
-            close={this.closeMainMenu.bind(this)}
-          />
-          <MainContainer>{children}</MainContainer>
-        </div>
-      </MuiThemeProvider>
-    );
-  }
-}
-
-class MainContainer extends React.Component {
-  render() {
-    const style = {
-      position: 'absolute',
-      left: 0,
-      top: 64,
-      right: 0,
-      bottom: 0,
-    };
-    return <main style={style}>{this.props.children}</main>;
-  }
-}
