--- src/redux/action/index.jsx
@@ -1,9 +1,0 @@
-export function timetable() {
-		return { type: 'timetable' };
-	}
-export function eventlist() {
-		return { type: 'eventlist' };
-	}
-export function course() {
-		return { type: 'course' };
-	}
