--- src/redux/store/index.jsx
@@ -1,7 +1,0 @@
-import { createStore } from 'redux';
-import reducer from '../reducers/';
-
-export default function configureStore() {
-	const store = createStore(reducer);
-	return store;
-}
